requirements.txt
django==2.2.1
django-cors-headers==3.0.2
django-cors-middleware==1.5.0
django-oauth-toolkit==1.2.0
django-filter==2.1.0
django-compat==1.0.15
djangorestframework==3.9.4
django-background-tasks==1.2.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
django-sslserver==0.20
msgpack==0.5.6
thinc==6.12.1
ecdsa==0.13
future==0.17.1
idna==2.8
kombu==4.2.1
Naked==0.1.31
opentok==2.10.0
pyasn1==0.4.4
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.6.4
python-jose==3.0.1
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rsa==4.0
shellescape==3.4.1
six==1.12.0
urllib3==1.24.1
vine==1.1.4
uuid==1.30
gunicorn==19.9.0
sendgrid==6.2.0
